pyproject.toml
[tool.poetry]
name = "fastapitemplate"
version = "0.1.0"
description = ""
authors = ["Stanislav Lukyanov <some email>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
black = "^24.4.2"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
redis = "^5.0.6"
celery = "^5.4.0"
flower = "^2.0.1"
fastapi-cache2 = "^0.2.1"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"

[tool.pytest.ini_options]
pythonpath = [
    "."
]
filterwarnings = [
    "ignore::DeprecationWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
asyncio_mode="auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
